Expose sanityCheck to integrationTest and integrationExternalTest
