+++ Heroku/backend/feed/src/integrationTest/kotlin/net/twisterrob/cinema/cineworld/sync/syndication/FeedData.kt
@@ -1,6 +1,5 @@
 package net.twisterrob.cinema.cineworld.sync.syndication
 
-import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Assertions.assertTrue
 
 object FeedData {
@@ -70,22 +69,4 @@ fun Feed.verifyHasAllAttributes(attributes: Set<Feed.Attribute>) {
 			"$attribute not found in ${this.attributes}"
 		}
 	}
-}
-
-fun Feed.sanityCheck() {
-	verifyAllAttributesAreValid()
-	verifyCalculatedProperties()
-}
-
-fun Feed.verifyAllAttributesAreValid() {
-	this.attributes.forEach { attribute ->
-		assertNotNull(attribute.code)
-		assertNotNull(attribute.title)
-	}
-}
-
-fun Feed.verifyCalculatedProperties() {
-	this.films.forEach { it.attributeList }
-	this.cinemas.forEach { it.serviceList }
-	this.performances.forEach { it.attributeList }
 }
